refactor: extract out password input
